% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_codons.R
\name{create_codons}
\alias{create_codons}
\title{Create codons from bases}
\usage{
create_codons(bases, start = 1)
}
\arguments{
\item{bases}{List of DNA bases.}

\item{start}{The wanted start base. The default value is 1.}
}
\value{
The codons coded for by the input bases.
}
\description{
Create codons from bases
}
