% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNA_creator.R
\name{DNA_creator}
\alias{DNA_creator}
\title{Randomized DNA Creator}
\usage{
DNA_creator(DNA_len)
}
\arguments{
\item{DNA_len}{The length of the wanted output DNA string}
}
\value{
A DNA sequence of length DNA_len
}
\description{
Randomized DNA Creator
}
\examples{
DNA_creator(5)
}
