% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_to_rna.R
\name{dna_to_rna}
\alias{dna_to_rna}
\title{dna transcription}
\usage{
dna_to_rna(dna_sequence)
}
\arguments{
\item{dna_sequence}{a sequence of dna}
}
\value{
the rna transcribed version of the sequence
}
\description{
dna transcription
}
