% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_codons.R
\name{extract_codons}
\alias{extract_codons}
\title{Extract Codons}
\usage{
extract_codons(sequence, start = 1)
}
\arguments{
\item{sequence}{RNA string that you wish to extract codons from, char}

\item{start}{Starting point of codon extractio, default at position 1}
}
\value{
substrings of three characters (one codon) in vector
}
\description{
Extract Codons
}
\examples{
extract_codons("AUGGCUUAA")
}
