% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_base_counts.R
\name{plot_base_counts}
\alias{plot_base_counts}
\title{Plot sequence base count}
\usage{
plot_base_counts(sequence)
}
\arguments{
\item{sequence}{RNA string that you which to plot base frequencies of}
}
\value{
a bar plot of frequencies of bases
}
\description{
Plot sequence base count
}
\examples{
plot_base_counts("AUGGCUUAA")
}
