% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_codons.R
\name{plot_codons}
\alias{plot_codons}
\title{Plots the amino acid distributions of a string}
\usage{
plot_codons(amino_acids)
}
\arguments{
\item{amino_acids}{String of amino acids to plot}
}
\value{
Returns a codon column plot object
}
\description{
Plots the amino acid distributions of a string
}
