% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA_to_aa_trans.R
\name{RNA_to_aa_trans}
\alias{RNA_to_aa_trans}
\title{RNA to amino acid translator}
\usage{
RNA_to_aa_trans(codons)
}
\arguments{
\item{codons}{a list of aminao RNA codons.}
}
\value{
a concatenated string of the amino acids names abbreviations.
}
\description{
RNA to amino acid translator
}
\examples{
Amino_acids <- RNA_to_aa_trans(c("UUU","ACC"))
Amino_acids

}
