% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_nucleotides.R
\name{sample_nucleotides}
\alias{sample_nucleotides}
\title{sample_nucleotides}
\usage{
sample_nucleotides(sequence_length)
}
\arguments{
\item{sequence_length}{define the length of wanted nucleotide sequence}
}
\value{
A nucleotide sequence based on the length provided "sequence length"
}
\description{
sample_nucleotides
}
