% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Translate to amino acids}
\usage{
translate(codons)
}
\arguments{
\item{codons}{List of codons}
}
\value{
One letter code corresponding to amino acids
}
\description{
Translate to amino acids
}
